requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
asteroid-filterbanks==0.4.0
audioread==3.0.1
bitarray==2.9.2
chainer==7.8.1
chumpy==0.70
ci-sdr==0.0.2
configargparse==1.7
crcmod==1.7
ctc-segmentation==1.7.4
cython==3.0.8
datasets==2.17.1
decorator==5.1.1
distance==0.1.3
editdistance==0.6.2
einops==0.7.0
espnet-tts-frontend==0.0.3
face-alignment==1.4.1
fairseq==0.12.2
fast-bss-eval==0.1.3
freetype-py==2.4.0
fvcore==0.1.5.post20221221
g2p-en==2.1.0
gast==0.5.4
h5py==3.10.0
humanfriendly==10.0
hydra-core==1.0.7
imageio==2.19.3
imageio-ffmpeg==0.4.7
inflect==7.0.0
iopath==0.1.10
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
jmespath==0.10.0
kaldiio==2.18.0
lazy-loader==0.3
librosa==0.9.2
llvmlite==0.41.1
lxml==5.1.0
modelscope==1.12.0
msgpack==1.0.7
nltk==3.8.1
numba==0.58.1
numpy==1.23.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
oss2==2.18.4
platformdirs==4.1.0
pooch==1.8.0
portalocker==2.8.2
pyarrow==15.0.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-core==2.14.6
pyglet==2.0.10
pynndescent==0.5.11
PyOpenGL==3.1.0
pypinyin==0.44.0
pyrender==0.1.45
pytorch-wpe==0.0.1
pyworld==0.3.4
pydub==0.25.1
resampy==0.4.2
sacrebleu==2.4.0
scikit-learn==1.3.2
sentencepiece==0.1.97
simplejson==3.19.2
smplx==0.1.28
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.2.0
timm==0.9.16
torch-complex==0.4.3
trimesh==4.0.8
typeguard==2.13.3
umap-learn==0.5.5
unidecode==1.3.8
yacs==0.1.8
yapf==0.40.2
transformers==4.24.0
funasr==1.1.16

python-multipart
fastapi[all]